SPLICE-1479 Make stats run on the OlapServer (2.5) (#724)

+++ hbase_sql/src/main/java/com/splicemachine/stream/KryoEncoder.java
@@ -52,7 +52,7 @@ public class KryoEncoder extends MessageToByteEncoder<Object> {
 
         output.flush();
         byte[] outArray = outStream.toByteArray();
-        out.writeShort(outArray.length);
+        out.writeInt(outArray.length);
         out.writeBytes(outArray);
     }
